fog/requirements.txt
numpy
scipy
mne
pyyaml
matplotlib
pyEDFlib
